pyproject.toml
[project]
name = "bela"
version = "0.1.0"
description = "Bela - a personal AI assistant"
authors = [{ name = "Deepjyoti Mukherjee", email = "djmsuman@gmail.com" }]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
  "celery (>=5.6.2,<6.0.0)",
  "django (>=6.0,<7.0)",
  "django-celery-beat @ git+https://github.com/celery/django-celery-beat",
  "django-celery-results (>=2.6.0,<3.0.0)",
  "djangorestframework (>=3.16.1,<4.0.0)",
  "psycopg (>=3.3.2,<4.0.0)",
  "psycopg-binary (>=3.3.2,<4.0.0)",
  "redis (>=7.1.0,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = ["ruff (>=0.14.11,<0.15.0)"]
